Move presidents store
